Make Foundations Definitions example easier to manipulate
When I went through the example, once I uncommented the lines I got errors from `cue`. The cause is pretty obvious -- the tutorial comments are not themselves commented -- but added friction in trying to learn what the tutorial was trying to convey.

I've added a second comment marker to ensure the tutorial comments remain commented even when the lines are uncommented to follow the Cuetorial.

I also realised that, to best demonstrate the effect, the second comment needed to be uncommented first. So I've added numbers to guide learners on what order to do things for best comprehension.