overview: remove mention of list arithmetic ops
The "*" and "+" operators haven't been mentioned in the CUE spec as
being able to apply to list operands for some while, and CUE v0.11 will
prevent them from working with lists.

This change stops mentioning these operators in the context of lists, in
the "overview" section. The rest of the site hasn't been exhaustively
reviewed to catch any other uses.

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>